% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\docType{class}
\name{bDif-class}
\alias{bDif}
\alias{bDif-class}
\title{An S4 class representing the Bayesian DIF/measurement invariance model}
\description{
Inherits from stanfit
}
\section{Slots}{

\describe{
\item{\code{data}}{The dataset provided when calling bDifFit.}

\item{\code{K}}{The number of latent groups estimated.}

\item{\code{model.type}}{The measurement model type (2PL currently supported)}

\item{\code{chain.max}}{The mcmc chain with the highest log posterior probability.}
}}

