% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\name{bDifFit}
\alias{bDifFit}
\title{Fit the Bayesian DIF/measurement invariance model}
\usage{
bDifFit(data, measurementModel, K, order, covariateModel = ~1,
  model.type = "2PL", method = "mcmc", ...)
}
\arguments{
\item{data}{A data.frame containing the indicators and any covariates.}

\item{measurementModel}{A RHS formula specifying the indicators for the latent factor.}

\item{K}{The number of latent groups to estimate.}

\item{order}{The item number whose difficulty (intercept) is ordered across latent groups.}

\item{covariateModel}{A RHS formula specifying the concomittant predictors of latent groups.}

\item{model.type}{A character string indicating the type of model to fit, either '2PL' or 'CFA'}

\item{method}{A character string indicating whether to use MCMC ('mcmc') or variational Bayes ('vb'). 'mcmc' is recommended.}

\item{...}{Arguments passed to \code{\link[rstan]{sampling}} or \code{\link[rstan]{vb}}}
}
\value{
bDif S4 object. See \link{bDif}
}
\description{
Fit the Bayesian DIF/measurement invariance model.
The model assumes that K latent groups exist between which some set of
items operate differentially.
The model also, currently, assumes a dichotomous state of DIF, 
such that DIF either exists for some item j or not.
Finally, one can predict latent group membership from known groups or other covariates.
The model jointly estimates item parameters, the probability of DIF for each item, 
the probability of each individual belonging to each latent group, and latent abilities.
}

