% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\docType{methods}
\name{clusters,bDif-method}
\alias{clusters,bDif-method}
\title{Obtain cluster membership assignments or probabilities for each individual.}
\usage{
\S4method{clusters}{bDif}(object, chains = object@chain.max, modal = TRUE)
}
\arguments{
\item{object}{A bDif object returned from \code{\link{bDifOut}}.}

\item{chains}{A numeric vector indicating the mcmc chain(s).}

\item{modal}{Logical. Whether to return the modal group membership vector or a matrix of cluster membership probabilities.}
}
\value{
Either a vector of length N containing modal cluster memberships (modal=TRUE) or an NxK matrix
  containing cluster membership probabilities.
}
\description{
By default, \code{clusters} will return the group each individual most probably belongs to.
If you want membership probabilities, change the \code{modal} argument.
}

