% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\docType{methods}
\name{coef,bDif-method}
\alias{coef,bDif-method}
\title{Obtain item parameter estimates}
\usage{
\S4method{coef}{bDif}(object, chains = object@chain.max, ltm = TRUE,
  cut = NULL)
}
\arguments{
\item{object}{A bDif object returned from \code{\link{bDifOut}}.}

\item{chains}{A numeric vector indicating the mcmc chain(s).}

\item{ltm}{Logical. Whether to include estimates from the ltm package.}

\item{cut}{Numeric between 0 and 1. If pp>50% is greater than this value, the non-DIF item parameters
are cut and the DIF item parameters retained. If lesser than this value, the opposite.}
}
\description{
Obtain item parameter estimates
}

