% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\name{factor.scores}
\alias{factor.scores}
\title{Compute factor scores for a latent variable model.}
\usage{
factor.scores(object, ...)
}
\arguments{
\item{object}{Object from a latent variable model model fit.}
}
\description{
Compute factor scores for a latent variable model.
}

