% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\name{loo.bDif}
\alias{loo.bDif}
\title{Leave-one-out cross-validation (LOO)}
\usage{
\method{loo}{bDif}(x)
}
\arguments{
\item{x}{bDif object from bDifFit.}
}
\description{
Efficient approximate leave-one-out cross-validation for Bayesian models.
}

