% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\docType{methods}
\name{posterior,bDif-method}
\alias{posterior,bDif-method}
\alias{posterior.bDif}
\title{Compute posterior probabilities of cluster membership.}
\usage{
\S4method{posterior}{bDif}(object, chains = object@chain.max, modal = FALSE)
}
\arguments{
\item{object}{bDif object from bDifFit.}

\item{chains}{Numeric vector. Specifies which mcmc chains to use for calculation.}

\item{modal}{Whether to compute posterior probabilities based on modal assignments (TRUE)
or to compute based on marginal probabilities.}
}
\description{
Computes posterior probabilities in one of two ways.
One (modal=FALSE) calculates the marginalized probability of cluster membership
across all cases.
The other (modal=TRUE) calculates the proportions of individuals in each cluster.
}

