% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\name{posterior}
\alias{posterior}
\title{Generic Method. Obtain posterior probabilities of cluster membership}
\usage{
posterior(object, ...)
}
\arguments{
\item{object}{Object containing cluster membership probabilities}

\item{...}{Further arguments for non-generic methods.}
}
\description{
Generic Method. Obtain posterior probabilities of cluster membership
}

