% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\name{waic.bDif}
\alias{waic.bDif}
\title{Widely applicable information criterion (WAIC)}
\usage{
\method{waic}{bDif}(x)
}
\arguments{
\item{x}{bDif object from bDifFit.}
}
\description{
Widely applicable information criterion (WAIC)
}

