% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\name{clusters}
\alias{clusters}
\title{clusters}
\usage{
clusters(object, ...)
}
\arguments{
\item{object}{Object containing cluster membership.}

\item{...}{Arguments passed to object methods.}
}
\description{
Generic function for extracting cluster memberships or probabilities.
}

