% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\docType{methods}
\name{coef,bDif-method}
\alias{coef,bDif-method}
\title{Obtain item parameter estimates}
\usage{
\S4method{coef}{bDif}(object, chains = object@chain.max, include = TRUE,
  cut = NULL)
}
\arguments{
\item{object}{A bDif object returned from \code{\link{bDifOut}}.}

\item{chains}{A numeric vector indicating the mcmc chain(s).}

\item{include}{Logical. Whether to include estimates from ltm (model.type=2PL) or lavaan (model.type=CFA)}

\item{cut}{Numeric between 0 and 1. If pp>50\% is greater than this value, the non-DIF item parameters
are cut and the DIF item parameters retained. If lesser than this value, the opposite.}
}
\value{
A matrix containing the relevant item parameter estimates:
\describe{
\item{Delta}{The mean posterior probability that the item exhibits DIF}
\item{pp>50\%}{The posterior probability that Delta is greater than 50\%}
\item{alpha.k}{The discrimination parameter for group k if DIF (model.type = '2PL')}
\item{diff.k}{The difficulty parameter for group k if DIF (model.type = '2PL')}
\item{int.k}{The intercept parameter for group k if DIF (model.type = 'CFA')}
\item{lam.k}{The factor loading parameter for group k if DIF (model.type = 'CFA')}
\item{res.k}{The residual variance for group k if DIF (model.type = 'CFA')}
}
}
\description{
Obtain item parameter estimates
}

