% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bDif.R
\docType{methods}
\name{factor.scores,bDif-method}
\alias{factor.scores,bDif-method}
\title{Compute factor scores for cases in bDif object.}
\usage{
\S4method{factor.scores}{bDif}(object, chains = object@chain.max)
}
\arguments{
\item{object}{bDif object from bDifFit.}

\item{chains}{Numeric vector. Which mcmc chain(s) to extract from.}
}
\value{
A vector of length N containing the posterior mean estimates of latent ability.
}
\description{
Compute factor scores for cases in bDif object.
}

